Code/requirements.txt
numpy==1.10.4
Pillow==3.1.1
cython==0.22  
chainer==1.5.1
nose==1.3.7
mock==1.3.0
sklearn==0.0
six==1.10.0
pyflakes==0.8.1
pep8==1.5.7
h5py==2.5.0
matplotlib==1.5.1
scikit-image
